+++ solarwinds_apm/oboe/transaction_name_pool.py
@@ -1,16 +1,6 @@
 import heapq
-import sys
 import time
 from functools import total_ordering
-
-if sys.version_info >= (3, 9):
-    # Use built-in generics for Python 3.9+
-    Dict = dict
-    List = list
-else:
-    # Use typing aliases for Python 3.8
-    from typing import Dict as dict
-    from typing import List as list
 
 
 @total_ordering