+++ tests/docker/install/install_tests.sh
@@ -82,7 +82,7 @@ function check_agent_startup(){
 }
 
 function install_test_app_dependencies(){
-    pip install flask requests
+    pip install flask requests psutil
     opentelemetry-bootstrap --action=install
 }
 
@@ -106,9 +106,28 @@ function run_instrumented_server_and_client(){
 
 
 # START TESTING ===========================================
-HOSTNAME=$(cat /etc/hostname)
-# docker-compose-set root of local solarwinds_apm package
-APM_ROOT='/code/python-solarwinds'
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    HOSTNAME=$(scutil --get LocalHostName)
+else
+    HOSTNAME=$(cat /etc/hostname)
+fi
+
+# Default to docker-compose-set root of local solarwinds_apm package
+if [ -z "$APM_ROOT" ]
+then
+    APM_ROOT="/code/python-solarwinds"
+    echo "Using default APM_ROOT: $APM_ROOT"
+else
+    echo "Using configured APM_ROOT: $APM_ROOT"
+fi
+
+if [[ "$MODE" == "local" ]]
+then
+    echo "Local mode: installing sdist and wheel locally"
+    pip install build
+    python -m build $APM_ROOT --sdist
+    pip -v wheel $APM_ROOT -w $APM_ROOT/dist --no-deps
+fi
 
 # Check sdist
 # shellcheck disable=SC1091